Notification Center is now supported on Mac OS X 10.8 and beyond. Growl is still supported and it receives notifications simultaneously; the user is therefore able to choose which notification scheme is preferable (and turn off MacTerm notifications in the other one).
This required significant extensions to the "Cocoa Future" module.
Notifications require calls to several classes that are not even
defined in older OS versions, and they require the definition of a
delegate that conforms to a new protocol.  Simple APIs were put in
place to allow notification code to only perform basic checks
("respondsToSelector:", etc.) while most of the trickery is handled
by the Cocoa Future module alone.

For simplicity the "Growl Support" module continues to be the only
place that notifications are actually sent.  The API magically
forwards the information it's been given to the system on 10.8+.
The API has also been slightly extended to support "subtitle" and
"sound" parameters, though no calling code has been changed yet to
take advantage of those settings.