Modified build of help content so that its tools and generated output are in the same place as other tools and generated output (and not in a subdirectory).
Makefiles for components will be relocated to the top
level with different names, instead of being in
subdirectories, so that it is easier to rely on shared
tools and content without resorting to parent-relative
pathnames.

Minor fix to HTMLHelpBuild.command to set a PATH that
includes the new default Xcode location.