Fixed the "Copy with Tab Substitution" command so that ranges of spaces are actually replaced by tabs!  The largest range of continuous spaces is user-defined but all smaller ranges are also iteratively replaced (e.g. if the user wants 8-space ranges replaced with tabs, this is done first; then, any remaining 7-space ranges become tabs; and so on).
Note that this type of substitution is internally handled
by certain text-copy routines in the Terminal module but
those were not used this time.  Instead of changing the
older code, Unicode-aware text-processing code in the
Terminal View layer was enhanced.  The copying routines
in the Terminal module (and indeed, copying of any kind)
will largely go away in favor of more efficient approaches
in the future.