Added the ability to search all open terminals for a phrase. Since the existing Find interface was already cramped, it was modified to move both the case-insensitivity option and the new "all terminals" option into a pop-up gear menu.
Fixed some issues with the display of status messages during
searches; in a few cases it appeared that a phrase was "not
found" when in reality the search just hadn't happened yet.

Exposed the entire terminal window list in Session Factory
as a short-cut.  This API can obviously change if the list
storage changes but it's far simpler to expose the list (and
iterate in a normal way) than to use callbacks or other
mechanisms.