Implemented the Cocoa version of the "Sessions" preferences panel, with initial support for only the "Keyboard" tab. As with other Cocoa-based preference panels, this version now supports "Inherit" checkboxes for all settings (old Carbon interfaces only had these in the "Formats" panel).
Since the Keyboard tab traditionally overloaded the behavior
of the "Control Keys" palette to serve as an input mechanism
for key mappings, the keypad can now send messages to Cocoa
objects.

Fixed some minor issues in the Preferences module and added
warning comments in Preferences_ContextGetData(): for the
moment, the "isDefault" flag returned by that routine will
only be set to true if you pass in the default context from
the preferences class of the requested setting (whereas in
reality, any default context successfully retrieves any
setting).  This is a potentially unreasonable burden of
consistency on calling code that may be revisited later.