Enhanced the Keypads module to be more explicit about event targets (Carbon) or responders (Cocoa) so that a module can only uninstall its own target or responder and no other.  The palette is now hidden automatically when a target or responder is uninstalled if it was only displayed because of the binding.  (If however the user originally displayed the palette, the palette remains displayed.)
Modified the Keyboard preferences pane (both the old
Carbon and new Cocoa versions) so the "Control Keys"
palette behaves more logically, e.g. automatically
unbinding from key-setting buttons in response to
actions such as closing the Preferences window or
switching panels.

Also made minor fixes to the kinds of panel messages
that are sent by the Preferences window so that the
"Control Keys" palette can be hidden in response to
those messages.