Ported contextual-menu code to Cocoa, and significantly simplified it in the process (for instance, removing several types of menus that were no longer used).  There is no longer a module for building contextual menus; instead, the code for building the menu is in Terminal View and the responsibility for constructing suitable NSMenuItem instances is in the Commands module.  The Context-Sensitive Menu module is now almost unnecessary but it does provide item-group management (automatically inserting separators as needed).
There is still a bit of a hack present to allow Carbon concepts
such as "command IDs" to work in a Cocoa environment.  This is
kept for historical reasons but also because in cases like this it
is actually far more convenient to have a command ID available.

Took the opportunity to add key equivalents to contextual menus.
This way, the user can use the keyboard to activate commands
while the menu is open (and learn what the key equivalents are).

Fixed some minor holes in contextual-menu logic.  For instance,
"Stop Speaking" is now available whether or not there is a text
selection, as long as speech is in progress; and, a selected
URL no longer hides other selection-related commands.