Added a preference to set the background color of the terminal cursor, and a setting to preserve the original "automatic" behavior.
Updated default format settings (used only when creating
new user preferences) to include cursor colors.  These are
not enabled by default.

Modified the blinking animation of the cursor so that it
is much closer to the scale of blinking text.  Previously
the animation used the timer from blinking text but it
adjusted the alpha value by the same amount per frame;
this looked unnatural and tended to make the cursor fade
away very quickly.

Both the Cocoa and Carbon interfaces have been updated to
support the new setting, though this is becoming quite
tedious; a permanent switch to Cocoa is overdue.