Major update; finally migrated the primary Xcode project to the Xcode 5.x format and the new LLVM compiler suite.
MacTerm now supports 10.9 minimum "in spirit" though the 10.6
SDK is currently still required even when building with Xcode 5.
(It must be manually installed to make this work.)  The reason
is the Carbon dependency: certain APIs such as QuickDraw are
not available in the 10.9 SDK.  Also, the 32-bit restriction on
most of Carbon means that 64-bit builds still aren't happening.
Obviously there has been lots of work to move away from Carbon
but until every last shred of Carbon is gone the build is
affected.

A very large number of code changes were made to deal with new
compilation warnings, the vast majority of which were to handle
unused return values consistently with a previously-defined macro.
Certain warnings still exist but they are relatively minor and
will be dealt with in a later update.

Fixed a couple of crashes that somehow only became obvious with
Mavericks: certain structures containing C++ classes were being
allocated with a raw allocator instead of operator "new".  This
could cause things like terminal window resizing and text search
to crash.

Determined why environment variable pollution was occurring in
only the first new shell window: the environment cleanup code
was not executing soon enough.  The cleanup phase is now one of
the first things that the front-end Python script does.

Removed various Pascal string utility routines and made minor
code changes to avoid depending on Pascal string utilities.