Removed a significant amount of code that is no longer used.
Removed some fairly straightforward dependencies on Carbon
APIs (that will not be available in later versions of the
Mac OS X SDK anyway).

As part of fixing API dependencies, changed the mechanism
used to choose a "Window menu rectangle" (for animation
purposes).  The new approach works correctly with a menu bar
that might be on a different screen in Mac OS X 10.9.