Initial migration of "generic dialogs" and (strongly related) "preference context dialogs" to Cocoa.  This first cut does little more than change the file type to Objective-C++ and add basic support for a parent window of type "NSWindow*".  The user interface is still Carbon-only.
These modules have traditionally allowed preference
panels to work the same in the Preferences window as
they do in sheets.  While this was very maintainable
for quite awhile, unfortunately panels had to "fork"
when it became clear that Carbon was going away and
that Carbon was not really compatible with Cocoa.
Preference panels therefore now have both Carbon and
Cocoa implementations that are nearly identical,
which (to put it mildly) is a maintenance headache.

A huge project milestone is to finally eliminate the
redundant Carbon versions of these user interfaces.
The Cocoa version of the Preferences window is very
close to completion (which is one half of the Carbon
dependencies).  The remaining panel dependencies
come from sheets that are presented by GenericDialog
and/or PrefsContextDialog.