Added a search interface to the new Cocoa-based Preferences window (this window is not yet enabled by default) and improved the way preferences panels are handled generally.
Search keywords are currently implemented as a property list inside the
application bundle, broken down by panel.  A partial match can occur at
the start of a word, or for a subset of words in a query string.

The results interface is just a menu with items for any panel that has
a matching keyword.  Since the matches can include individual tabs, the
action for each menu item has to switch to the right tab and ensure that
the tab's parent panel is visible.

This was therefore a good time to slightly refactor the way panels work
in Cocoa.  There is now a protocol that can be adopted by any class that
acts as a parent for panels (initially, only the tabs interface and the
Preferences window itself).  This allows panels to directly respond to
display requests by forwarding requests to parents.

The refactoring has also allowed some of the more tedious setup in the
Preferences window to be eliminated.  Now, panels can be installed from
a simple list of classes; all other configuration is automatic.