Fixed experimental Cocoa version of terminal view to use the correct colors for "search result" text.
Minor changes to the Cocoa terminal view code, such as adopting
"@property" syntax.  Also, the view now asks to be layer-backed.

At initialization time MacTerm is now forced to be the front
process if it is started while standard input is a terminal
(this aids debugging, especially for new features like the
Cocoa-based terminal view).