Created a foundation for calling XPC services, and an initial service (that does very little and can currently only be launched from the debug interface).  The first service is likely to be used for wrapping calls to Python callbacks, hence the name of the new skeleton service.
Note that the Cocoa Future module has been updated to support
calls to newer XPC service APIs so that it can be called from
a code base that is technically stuck on an older SDK version.
One of the goals of adopting XPC services is to allow some
project code to start migrating to newer SDKs, minimizing what
is affected by the time-consuming Carbon-to-Cocoa port.