Minor casting additions to fix method-match error.
It is possible for the compiler to complain about calls to
createImageWindowFrom() when the NSRect parameter comes
from a window content-view frame.  Since "contentView" in
NSWindow returns "id", the compiler may not know that the
"frame" method returns an NSRect type.  This was resolved
by explicitly casting the result of the "contentView" calls
to NSView* before calling "frame".