Cocoa Preferences improves animation, stability.
In the Cocoa-based Preferences window (currently
visible only in debug mode), the code for hiding
and showing the source list has been reworked to
use blocks and Core Animation.  This allows the
view changes to animate in sync with the window
and it allows the rate of change to be specified.

Other code changes: removed the window's Zoom
button (on El Capitán this causes problems because
the window is then allowed to occupy unsupported
sizes in split view), and fixed a possible crash.