Refactored glyph-drawing code to be range-based.
Although not necessarily obvious from the "diff", the large
"switch" in the initializer of the layer class has simply
been split and relocated to two different functions.  As
various Unicode ranges are implemented, they will be handled
in functions based on range instead of having a gigantic
switch to cover them all.

Also, removed an alternate initializer that was never used.