Cocoa Preferences now stays onscreen when resized.
In the Cocoa-based Preferences window (currently
visible only in debug mode), the window location is
now adjusted to keep the window entirely onscreen
after an automatic resize.  This behavior existed
for certain cases before but it has been extended
to work no matter which edge of the window is off
the screen.