Ported Panel object to use NSViewController.
This affects the vast majority of user interface objects.

This removes some NIB-handling code, and it allows the "view"
property to be read from the controller object.  Most other
benefits of NSViewController are not yet available because
the project uses the 10.6 SDK.

All NIBs that previously bound to the Panel "managedView"
outlet are now binding to the NSViewController "view" outlet.
For now, a wrapper method "managedView" remains in Panel as
an alias for the "view" property.