Terminals now support sub-parameters (with ":").
The terminal now interprets both ";" and ":" as parameter
delimiters.  For simplicity, all parameters are still kept
in one list and handlers still start with the first parameter
from a set; except, it is now possible for handlers to easily
identify sub-parameters.  Handlers that do not check for any
sub-parameters (most of them) will simply see all parameters
as a continuous sequence, which helps to avoid code changes.
This "equivalent by default" behavior will also make it
easier to be compatible with terminals and programs that
incorrectly use main parameters for values that belong in
sub-parameter sequences.