24-bit color support added and enabled by default.
Common standard and non-standard sequences for setting
arbitrary RGB color components are now recognized, and
terminal views can show up to 2048 additional colors
(NOT counting any references to the base 256 colors).
This is a very substantial and important enhancement
that allows for much more interesting terminal-based
application and game development!