Completed Cocoa panel for Workspaces preferences.
The mechanism for binding a collection of preferences was
extended to make it easy to include a Default list value.
The Workspaces interface also required more changes to the
bindings mechanism to correctly handle the varying "index"
in tags when accessing window-specific preference settings.
Added a Preferences_ContextIsDefault() API to make it
easier to adapt behavior based on use of Default.  And,
extended the Keypads module to provide a Cocoa-based way to
spawn the window-positioning interface.