Fixed a regression in displaying some preferences.
The PreferenceValue_Array and PreferenceValue_CollectionBinding
objects were calling the wrong initializer from the superclass,
and this was only evident after recent changes to the behavior of
the superclass (part of the Cocoa-based Workspaces panel code).
The "preferencesTag" property was not actually set for those two
cases.  This caused a zero to be read instead of the real setting
in some situations, and the user interface would display the data
that maps to zero; for instance, a Session Favorite could show a
"CR" line-ending setting (equal to zero internally) even if the
stored value was not "CR".  The initializer calls are now correct
and the PreferenceValue_InheritedSingleTag class now hides the
unsupported base-class initializer with an asserting version.