Completed Cocoa panel for Macros preferences.
This allows the user to edit macro sets, and it supports
inheritance (like all other Cocoa-based preference panels).
With this commit, there is now a Cocoa version for every
panel that is used by the Preferences window.  It will
therefore be possible to switch the default Preferences
command to use the Cocoa window in the near future,
eliminating a massive amount of redundant and obsolete
Carbon-based code.