Removed Carbon implementation of Formats panel.
Now that the Cocoa panel is the default, the code for the
Carbon version of the Formats panel is not needed at all.
Also removing the Color Box module, which was only needed
to implement the Carbon version of the Formats interface.
This removes hundreds of lines of code from the project!