Fixed potential crash in Custom New Session sheet.
The Custom New Session sheet has a special observer installed
for the embedded panel that was not removed, making it possible
for the system (at times) to crash when accessing the object.
The close-handler has been updated to remove the observer,
which required exposing the panel object from the Generic Dialog.
Also, relocated some initialization code in the embedded panel
(it wasn't having issues but it is more correctly placed in its
new location).