Added debug-logging mode for lock objects.
Pointer-lockers can now be told to write messages and
stack traces to the console when locks are acquired
or released.  This way, it is possible to see exactly
which execution paths cause the lock counts to change,
and keep all calls balanced.  Most modules use lockers
to control object deletion so this change will help to
prevent resource leaks.