Added “weak reference” implementation for C++.
Created a scheme by which an object can easily hold a reference
to another object, and have that reference auto-cleared when the
target is destroyed or otherwise invalidated.  This is largely
necessary due to dependencies on older compilers, and due to the
use of C++.

This change includes a unit test function for the new behavior,
and fixes typos in the comments of other unit test functions.

Also note, versions in new comment blocks now use "(YYYY.MM)"
date formats instead of trying to approximate release numbers.