Removed some dependencies on old APIs.
Further reduced dependencies on APIs that are not
available in the latest SDKs.  Some unused functions
were removed.  Certain applications of QuickDraw
regions have been reimplemented, which could help
slightly with performance.  And when a deprecated
API was only used for trivial data manipulation
(such as setting points or shifting rectangles),
an equivalent local function was created.