Moved Print Preview into a separate process.
The window that is displayed prior to printing text from
a terminal is now in an entirely separate process, for
added stability.  Even if MacTerm quits, preview windows
remain usable and they can even print successfully.  Also,
there is no longer any limit on previews; instead of being
a single application-modal dialog, multiple terminals can
spawn print jobs in parallel.  The sub-process is also
sandboxed, having only the basic entitlement of printing.

To give the illusion of seamless window cycling between
processes, a new “child process proxy window” concept has
been introduced.  This special window controller is aware
of a running sub-process (in this case, Print Preview),
and it responds to window-like actions by switching to
the sub-process as appropriate.  Currently, this means
that when the user cycles through windows, both MacTerm
and Print Preview windows are visited automatically even
though they belong to different processes!  In the future,
this may be extended to have additional features.