Fixed some sheet behavior problems.
It it now possible to click anywhere in a background
terminal window to bring it to the front, even if it
has a sheet.  Also, fixed a problem where event
handlers may force popovers to reappear instead of
respecting their invisible states (such as when the
vector graphics window’s “Rename…” command was used).
Fixed a small memory leak in code handling observers.