Converted Window Title Dialog into shared code.
Changed Window Title Dialog module so that it is
more generic and located in the Shared tree.  Any
elements of the API that were specific to Session
and Vector Window have been replaced by blocks.
The build has been modified so that XIB files can
also be read from the Shared tree.