Fixed possible exception during animation.
In certain cases, such as search results that zoom to
an offscreen location, the target rectangle can be
infinite and trigger an exception.  For now, this
case is detected and the animation window is just not
created.  Although, the search mechanism should be
revised to handle scrollback search results properly.