Added support for ".itermcolors" file imports.
It is now possible to import any number of files in the
".itermcolors" format, such as the vast number of color
schemes available online.  These can be read by using
the "+" menu of the Preferences window, or by dragging
files to the MacTerm Dock icon, or by using the Finder’s
"Open With" contextual menu.  MacTerm is able to copy
most colors directly, although certain values require
special interpretation as described in the release notes
and help pages.  Also made general improvements to the
error-handling capabilities of the Quills Python API for
opening files.