Enhanced network addresses window.
Significantly enhanced the "Network Addresses of This Mac"
window (previously named "IP Addresses of This Mac"): added
a toolbar with a refresh feature and progress indicator, and
a footer reminding the user that drag-and-drop is an option.
After seeing delays and empty results on macOS Sierra, the
mechanism for finding network addresses has been reimplemented
to use more modern interfaces, asynchronous blocks and caching.
The macro feature has been modified to use this new mechanism
and to have a built-in "retry" if an address is not initially
found.  There is a new "\I" macro sequence too, expanding into
the entire list of network addresses (for better flexibility).
Finally, fixed recent problems with editing user interfaces
from Xcode 8 by removing and adding back every XIB/NIB file
in the project.