Exported settings can now copy from Default.
Previously, exported files would perfectly preserve the
source context, meaning that they could only contain
settings that are not inherited.  This produced an odd
situation: if the file was imported much later, or
shared with another user, the source of inherited values
could be quite different and produce a final collection
that is undesirable.  A new option (enabled by default)
has been added to the save panel that allows the user to
include copies of current Default values for any missing
settings, producing a perfect snapshot of the collection
as seen by the exporting user.  Since this can produce a
larger file and is less flexible, the old behavior is
still available by flipping the checkbox.