Preferences tweaks for name keys.
The preference keys "name" and "name-string" are now
explicitly removed from global settings at startup
time if they exist (possibly due to a past import gone
wrong).  These values are only used to name separate
collections so they make no sense in global settings.
Also, an export of a Default set should not include
any name keys in the file, to avoid confusion and to
make sure that future imports of the data do not try
to use a name.