Fixed focus-ring rendering in more cases.
Extended the set of “core” UI classes that perform their
own focus-ring rendering in layer-backed view hierarchies,
fixing several more cases of missing or clipped focus rings
when using Full Keyboard Access.  This means that it is now
much easier to navigate various interface elements from the
keyboard (as it was in the past, prior to the introduction
of layers).  Also, since this required updating most XIB
files anyway, all XIBs have been changed to target the
“deployment target” OS version instead of using OS 10.6.
The property list minimum version is now also set to 10.10.