Commands for resize now adhere to screen edges.
The "Make Screen Wider" and "Make Screen Taller" commands
now examine the resulting window size and automatically
extend the window to the corresponding screen-edge if it
is nearly at the edge.  The effect of this is to extend
the matte region slightly and make it easier to align
windows cleanly on the screen.