Initial support for Touch Bar.
Added a foundation for various windows and views to
support Touch Bar items more easily (essentially by
creating a separate XIB that is loaded only when
Touch Bar is supported).  Created initial controllers
for Touch Bar items in Preferences, terminal windows,
Session Info and vector graphics windows.  So far
there are only two types of items: Search (works in
Preferences and terminal windows) and Full Screen
(works in terminals and vector graphics windows).
Since the project SDK is still at 10.6, several of
the Touch Bar mechanisms are wrapped in some way.
Also, it appears that the customization interface is
not reliable so it is not yet enabled; perhaps it
can be enabled after moving to a later SDK version.