Changed feedback E-mail to use Bug Reporter.
The MacTerm Bug Reporter helper application is now used to
compose E-mail when “Provide MacTerm Feedback” is selected
from the MacTerm application menu.  This allows the new
message to automatically include details on the OS version
and application version.  The Bug Reporter has also been
sandboxed now, with no entitlements.  It has significantly
changed to recognize environment variables that configure
behavior: for instance, whether or not to display a message
about a bug, the type of bug (crash or not), and ultimately
the content of any E-mail message that is composed.  The
code has also been tweaked to use more Cocoa, which makes
some things simpler.