Major improvements to the floating Command Line.
A number of “shell-like” key equivalents have been added to the
floating Command Line (now documented in the Help pages).  For
example, control-L and Escape are now sent immediately; partial
command strings can be sent with Tab or control-D; and, Delete
key-presses cause a backspace or delete character to be sent if
the command-line field is empty.  The window now also displays
any partially-complete text that has been sent (by a Tab, say).
Fixed the Command Line so that a simple Return when the field is
empty will still cause a new-line to be sent, instead of nothing.
The command-Return key equivalent can now be used to send text
immediately without automatically adding a new-line at the end.
Finally, the Command Line now uses more Default preferences, like
the font, the cursor color and the colors for any selected text.