Default button setting is delayed in some cases.
In order to prevent users from accidentally dismissing alerts
that appear automatically, the default button may now be set
with a delay.  This is also automatic for any dialog that has
only one action button, like the message that appears in
terminals if there are too many text encoding errors (a simple
message with one “OK” button).  If a user happens to type a
Return key into a terminal as soon as such a message appears,
the first key-press will now be rejected.  Also, until the
short delay is over, the action button will have a normal
appearance; it will only be highlighted when the Return key
mapping is set.  This delay is not used with most alerts and
dialogs; it is currently only used for dialogs that appear
without a direct action, such as the text encoding error.