Added more precise data type for pixel values.
Modified certain field values and APIs so that pixel widths and
heights in terminal views allow both precise (floating-point)
and integral (legacy) values, depending on the drawing target.
Similarly, font sizes can now be specified as floating-point.
The distinction is handled by a new wrapper class with methods.
High precision is necessary for correct layout calculations in
new code that uses Core Graphics in a Cocoa view, since pixels
are not as constrained.  Until the legacy QuickDraw-based
renderer can be retired though, it is still necessary to have
the option of integer calculations.