Fixed terminal selection and scrolling issues.
Corrected recent regression when clicking in the scrollback
region, as selected text would be off by one line from the
click location.  Also, auto-scroll now works properly
(triggered when text selections are extended to locations
that are outside the visible terminal screen), including
drags off the top or bottom of the screen, and drags where
the mouse has reached the edge of the display.  A console
debugging message for out-of-bounds has also been disabled
because it would trigger frequently during auto-scroll.