Substantial accessibility improvements.
The vast majority of user interface elements have
been extended in some way to have better accessibility
descriptions, which greatly improves the behavior of
assistive technologies such as Voice Over.  In most
cases, this only required extra context (for example,
to point out the group that a control applies to,
because the title alone is vague).  Some controls had
no identification at all!  The most extensive change
was to add accessible descriptions to every check box
for returning settings to Default values, because
these have no labels to describe the settings that
they affect.