Added capabilities to debug commands.
All debugging scripts now depend on a single Python program to
operate.  The Python program is capable of running MacTerm or any
internal applications installed in the main bundle.  The script
automatically determines if it should attach to a running process
or launch a new process.  Also, before attaching to a process,
the paths of all candidates are checked to ensure that the target
comes from the same tree that launched the debugging script (this
allows one MacTerm shell to attach to a completely different copy
of MacTerm, and not accidentally attach to itself).  The script
also knows about internal applications that are dependent on the
environment, which causes the debugging session to fail when no
running process is found (instead of launching something new).
Finally, the script is compatible with both Python 2.x and 3.x.