Added menu short-cuts to Preferences toolbar.
Preferences window toolbar icons now display menu
buttons for categories that have sub-panels (tabs).
If the icon portion is clicked, the last-used tab
is displayed, as before; otherwise, a menu of all
tabs is shown and the user can navigate directly.
The Full Screen panel has increased its minimum
width to ensure that all toolbar icons are still
visible, since most toolbar icons are a bit wider.
This also required extending the Panel interface to
give a hint about the number of children in a panel
(which is why source files have changed that are
not related to the Preferences window); the menu
short-cut is only displayed if there are at least
two child panels available.