Added auto-scroll to macro-based text search.
Macros with an Action of "Find in Local Terminal Verbatim" or
"Find in Local Terminal with Substitutions" will now trigger
automatic scrolling to reveal the matching line, just like a
normal search would do.  This makes macros much more useful, as
they can be used to jump back to a certain point in the buffer
based on a pattern match.