Bug fixes and enhancements for Undo/Redo.
Menu command handlers now check Cocoa windows for possible
NSUndoManager instances, deferring to those if they exist
(this allows pure-Cocoa windows such as the Command Line to
properly handle Undo and Redo for text editing actions).
Also, Undo and Redo are now supported if a sheet is used to
modify the dimensions of a terminal screen.