Fixed initializers of ChildProcessWC_Object.
Added missing initializers to ChildProcessWC_Object that were
reported after testing on a future compiler.  Removed the NIB
dependency and refactored the controller so that it can call
the designated initializer of the superclass.  Also fixed the
offscreen origin point to fall within the limits that are
documented for the window server.