Implemented base-class designated initializers.
Fixed several cases of Objective-C subclasses not implementing
the designated initializers of their base classes (which newer
Objective-C compilers will warn about), and fixed a couple of
other warnings in the affected files.  Most of these overriding
initializers are not valid, triggering assertion failures; a
new preprocessor macro is used to mark invalid initializers
that are only present to satisfy the compiler.