Fixed numbered-list selections after sorting.
The Macros and Workspaces preferences panels now show
the correct panel for the selected item, even after
the numbered list has been sorted differently.  (That
is, the display is now in sync with the item actually
selected in the list, and not the absolute position of
the selected row in the list.)  Also, the corresponding
properties for binding to list items have been renamed
so that they are not confused with the actual headers
of table views.