Initial commit of some Unicode string utilities.
This adds a stable subset of new Unicode functionality that is
being actively developed.  It includes a new unit test driver
for the String Utilities module, a renamed Objective-C source
file, and helper functions for iterating over strings and
analyzing strings.  In this first set of changes, new functions
are called only by unit tests.