Changed main application icon artwork and format.
The application icon is now generated from the newer
".iconset" file format, indirectly through a script
that generates a collection of different-sized PNGs.
This means that the new ".icns" in the bundle supports
much higher resolutions than before, and Retina sizes.
The icon also looks different: it is now wider (using
a golden-ratio rectangle) and it has a different
border and shadow.  The two session status icons have
also changed to use the same appearance for their
terminal screens.