Fixed animations on macOS High Sierra.
Removed code that was directly setting the image of a layer that
was associated with a view, which is unsupported behavior (and
the images were not rendered at all on macOS High Sierra).  The
alternate code that used NSImageView is now run all the time.
Also, consolidated the image-window-creation functions so that
the window is only created in the version that receives an image
(the 2nd function delegates to the basic image window function).