Certain macros are now enabled all the time.
When a macro has the "Open URL", "New Window with Command" or
"Select Window by Title" actions, it is now available all the
time in the Macros menu instead of requiring an active
terminal window.  (The code has been slightly tweaked to make
sure that such macros work when no session is present.)  In
addition, all the commands that switch between macro sets are
now enabled all the time.