Menu with macros now display inherited macros.
The main Macros menu and terminal contextual menus (with text
selected) will now display macros inherited from Default, when
those macros have not been reimplemented by the active set.
This was technically always the behavior of the main Macros
menu but the inherited actions had blank names.  This is new
behavior for contextual menus.  Documentation has been updated
to make this clear.  This provides a way for users to mix
macros that should be available more often, with specialized
macros that are only available in a particular set.