Basic Sixel graphics decoding.
Initial support for the VT340 Sixel graphics format, with
a “low level” parser of all necessary terminal sequences
and Sixel graphics data.  A simple preference setting has
been defined so that the Sixel decoder can be disabled if
necessary.  This first commit does not include any way to
render the graphics; the parsed graphics sequences are
simply logged for now.