Extended terminal attributes to hold bitmaps.
This changes only the low-level definition of terminal
text attributes, adding new bits that determine how
bitmap image sections may be associated with particular
cells.  This definition will make it easy for terminal
rendering code to find what it needs, and it will allow
other requirements to be met (such as bitmaps that can
scroll inline or be partially replaced).  The view has
not been updated yet for rendering images, and the
Sixel decoder has not been extended yet to use any of
these new attribute bits.