Moved mostly away from QuickDraw regions.
Removed most data structures and code that dealt with legacy
QuickDraw “regions”, updating some of the dependent code to
use Core Graphics in the process.  The only cases left that
require QuickDraw regions are in Carbon views: drag-and-drop
for Carbon has no other option, and region queries in custom
views technically must still support QuickDraw regions even
though they now also support the more precise “shape” type.
This was a difficult change to do correctly without creating
visual artifacts, and was simplified by forcing the Carbon
views to now be completely opaque.  (Text overlaid on images
or other backgrounds will be revisited in the Cocoa view.)

Some graphics debugging features have been added as well
(disabled by default).  One change is to have a way to log
the coordinates of every rectangle in an updated region.
Another enhancement in both the Carbon and Cocoa views is
the ability to color the given update region, making it
easier to see when and where updates occur.  The background
also rotates through a few different colors so that updates
are clear even if they end up in adjacent regions.