Enhancements and fixes to new SIXEL decoder.
Fixed Sixel decoder to also return the graphics cursor to
the left margin when a new line is encountered, per spec.
The decoder structure now keeps track of more information,
including any suggested image width/height from the header
and the maximum distance that the graphics cursor moves in
both directions.  (This will later help with calculating
the portion of the terminal that is covered by an image.)
Also added methods to help with aspect ratio calculations.